src/api/requirements.txt
fastapi[standard]
uvicorn
httpx
semantic-kernel[azure,mcp]>=1.27.2
azure-appconfiguration>=1.4.0
azure-identity>=1.14.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
azure-monitor-opentelemetry
azure-monitor-opentelemetry-exporter>=1.0.0b35
python-dotenv>=1.0.0
redis
